Refs #4598. Converting separated files into cpp.
