--- test/blackbox/BlackboxTestsUtils.hpp
@@ -1,32 +1,0 @@
-// Copyright 2019 Proyectos y Sistemas de Mantenimiento SL (eProsima).
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-uint16_t global_port = 0;
-
-uint16_t get_port()
-{
-    uint16_t port = static_cast<uint16_t>(GET_PID());
-
-    if (5000 > port)
-    {
-        port += 5000;
-    }
-
-    std::cout << "Generating port " << port << std::endl;
-    return port;
-}
-
-#include "utils/print_functions.hpp"
-#include "utils/data_generators.hpp"
-#include "utils/lambda_functions.hpp"
